+++ tests/testthat/test-0-test-package-installation.R
@@ -1,12 +1,9 @@
-context("Installation of Test Packages")
-rm(list = ls())
-
 # Test packages are installed in a temporary directory before testing and
 # uninstalled after. Many pkgnet tests depend on these packages installing
 # correctly and being available.
 
 test_that('Test packages installed correctly',{
-    testPkgNames <- c("baseballstats", "sartre", "milne", "silverstein")
+    testPkgNames <- c("baseballstats", "sartre", "milne", "silverstein","control")
     for (thisTestPkg in testPkgNames) {
         expect_true(
             object = require(thisTestPkg