+++ tests/testthat/test-DependencyReporter-network.R
@@ -1,5 +1,3 @@
-context("DependencyReporter Expected Network Tests")
-
 ##### TESTS #####
 
 test_that('DependencyReporter extracts expected network for baseballstats', {
@@ -23,3 +21,23 @@ test_that('DependencyReporter extracts expected network for baseballstats', {
     )
 })
 
+test_that('DependencyReporter extracts expected network for milne (with LinkingTo:)', {
+  
+  testObj <- DependencyReporter$new()$set_package('milne')
+  
+  # Test nodes
+  expect_equivalent(
+    object = testObj$nodes
+    , expected = data.table::fread(file.path('testdata', 'milne_dependency_nodes.csv'))
+    , ignore.col.order = TRUE
+    , ignore.row.order = TRUE
+  )
+  
+  # Test edges
+  expect_equivalent(
+    object = testObj$edges
+    , expected = data.table::fread(file.path('testdata', 'milne_dependency_edges.csv'))
+    , ignore.col.order = TRUE
+    , ignore.row.order = TRUE
+  )
+})